requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
st-gsheets-connection>=0.0.3
plotly>=5.0.0
google-api-python-client>=2.0.0
cloudinary>=1.36.0
docxtpl>=0.16.0
docx2pdf>=0.1.8
bcrypt>=4.1.0

openpyxl>=3.1.0
google-generativeai>=0.3.0
